Further cleanup and simplification
